refactor(layout): decompose artifact layout
