--- lib/layout/artifacts/ArtifactLayout.js
@@ -1,1175 +1,0 @@
-import { is, getDefaultSize } from '../../di/DiUtil.js';
-import { LayoutError } from '../../LayoutError.js';
-import {
-  ROUTING_MARGIN,
-  VERTICAL_GAP,
-  PARTICIPANT_HEADER_WIDTH,
-  ANNOTATION_MIN_WIDTH,
-  ANNOTATION_MAX_WIDTH,
-  ANNOTATION_WIDTH_STEP,
-  ANNOTATION_CHARACTER_WIDTH,
-  ANNOTATION_LINE_HEIGHT,
-  ANNOTATION_PADDING,
-  ANNOTATION_MIN_HEIGHT,
-  ANNOTATION_TARGET_ASPECT_RATIO,
-  ANNOTATION_ASPECT_RATIO_PENALTY_SCALE,
-  MAX_ARTIFACT_SEARCH_OFFSET,
-  MAX_ARTIFACT_GAP_STEPS,
-  NON_STRAIGHT_ARTIFACT_ASSOCIATION_PENALTY,
-  BOUNDARY_EVENT_ARTIFACT_CLEARANCE
-} from '../Constants.js';
-import {
-  point,
-  bounds,
-  rectanglesOverlap,
-  routeLength,
-  toSegments,
-  segmentEntersRect,
-  segmentsProperlyCross,
-  compareScores
-} from '../geometry/index.js';
-import { pathIsClear } from '../process/routing/SequenceFlowRouting.js';
-
-const ASSOCIATION_OBSTACLE_CLEARANCE = 1;
-
-export function findContainingArtifactContainers(elementBounds, shapes) {
-  const center = {
-    x: elementBounds.x + elementBounds.width / 2,
-    y: elementBounds.y + elementBounds.height / 2
-  };
-
-  return [ ...shapes.entries() ]
-    .filter(([ element, rect ]) => {
-      return (
-        is(element, 'bpmn:Lane') ||
-        is(element, 'bpmn:Participant') ||
-        is(element, 'bpmn:SubProcess')
-      ) &&
-      center.x >= rect.x &&
-      center.x <= rect.x + rect.width &&
-      center.y >= rect.y &&
-      center.y <= rect.y + rect.height;
-    })
-    .map(([ , rect ]) => rect)
-    .sort((a, b) => a.width * a.height - b.width * b.height);
-}
-
-function associationConnection(
-    association,
-    owner,
-    ownerBounds,
-    artifact,
-    artifactBounds,
-    connectionIndex = 0,
-    connectionCount = 1,
-    obstacles = [],
-    routeSegments = []) {
-  const artifactIsSource = is(association, 'bpmn:DataInputAssociation') ||
-    association.sourceRef === artifact ||
-    (Array.isArray(association.sourceRef) && association.sourceRef.includes(artifact));
-  const dataArtifact = is(artifact, 'bpmn:DataObjectReference') ||
-    is(artifact, 'bpmn:DataStoreReference');
-
-  if (dataArtifact) {
-    const points = orthogonalAssociationRoute(
-      owner,
-      ownerBounds,
-      artifact,
-      artifactBounds,
-      obstacles,
-      routeSegments,
-      connectionIndex,
-      connectionCount
-    );
-
-    return artifactIsSource ? points : points.reverse();
-  }
-
-  let ownerPoint;
-  let artifactPoint;
-
-  if (artifactBounds.y + artifactBounds.height <= ownerBounds.y) {
-    const offset = associationDockOffset(
-      ownerBounds.width,
-      artifactBounds.width,
-      connectionIndex,
-      connectionCount
-    );
-    ownerPoint = point(ownerBounds.x + ownerBounds.width / 2 + offset, ownerBounds.y);
-    artifactPoint = point(
-      artifactBounds.x + artifactBounds.width / 2 + offset,
-      artifactBounds.y + artifactBounds.height
-    );
-  } else if (ownerBounds.y + ownerBounds.height <= artifactBounds.y) {
-    const offset = associationDockOffset(
-      ownerBounds.width,
-      artifactBounds.width,
-      connectionIndex,
-      connectionCount
-    );
-    ownerPoint = point(
-      ownerBounds.x + ownerBounds.width / 2 + offset,
-      ownerBounds.y + ownerBounds.height
-    );
-    artifactPoint = point(
-      artifactBounds.x + artifactBounds.width / 2 + offset,
-      artifactBounds.y
-    );
-  } else if (artifactBounds.x + artifactBounds.width <= ownerBounds.x) {
-    const offset = associationDockOffset(
-      ownerBounds.height,
-      artifactBounds.height,
-      connectionIndex,
-      connectionCount
-    );
-    ownerPoint = point(ownerBounds.x, ownerBounds.y + ownerBounds.height / 2 + offset);
-    artifactPoint = point(
-      artifactBounds.x + artifactBounds.width,
-      artifactBounds.y + artifactBounds.height / 2 + offset
-    );
-  } else if (ownerBounds.x + ownerBounds.width <= artifactBounds.x) {
-    const offset = associationDockOffset(
-      ownerBounds.height,
-      artifactBounds.height,
-      connectionIndex,
-      connectionCount
-    );
-    ownerPoint = point(
-      ownerBounds.x + ownerBounds.width,
-      ownerBounds.y + ownerBounds.height / 2 + offset
-    );
-    artifactPoint = point(
-      artifactBounds.x,
-      artifactBounds.y + artifactBounds.height / 2 + offset
-    );
-  } else {
-    const artifactAbove = artifactBounds.y + artifactBounds.height / 2 <
-      ownerBounds.y + ownerBounds.height / 2;
-    const offset = associationDockOffset(
-      ownerBounds.width,
-      artifactBounds.width,
-      connectionIndex,
-      connectionCount
-    );
-    ownerPoint = point(
-      ownerBounds.x + ownerBounds.width / 2 + offset,
-      artifactAbove ? ownerBounds.y : ownerBounds.y + ownerBounds.height
-    );
-    artifactPoint = point(
-      artifactBounds.x + artifactBounds.width / 2 + offset,
-      artifactAbove ? artifactBounds.y + artifactBounds.height : artifactBounds.y
-    );
-  }
-
-  return artifactIsSource
-    ? [ artifactPoint, ownerPoint ]
-    : [ ownerPoint, artifactPoint ];
-}
-
-function associationDockOffset(ownerSpan, artifactSpan, index, count) {
-  if (count < 2) {
-    return 0;
-  }
-
-  const span = Math.min(ownerSpan, artifactSpan);
-
-  return (index + 1) * span / (count + 1) - span / 2;
-}
-
-function orthogonalAssociationRoute(
-    owner,
-    ownerBounds,
-    artifact,
-    artifactBounds,
-    obstacles,
-    routeSegments,
-    connectionIndex,
-    connectionCount) {
-  const horizontal = horizontalAssociationDocks(
-    artifactBounds,
-    ownerBounds,
-    connectionIndex,
-    connectionCount
-  );
-  const vertical = verticalAssociationDocks(
-    artifactBounds,
-    ownerBounds,
-    connectionIndex,
-    connectionCount
-  );
-  const centeredIndex = connectionIndex - (connectionCount - 1) / 2;
-  const channelOffset = centeredIndex * ROUTING_MARGIN;
-  const candidates = [
-    {
-      points: [
-        horizontal.artifact,
-        point(vertical.owner.x, horizontal.artifact.y),
-        vertical.owner
-      ],
-      artifactAxis: 'horizontal',
-      ownerAxis: 'vertical'
-    },
-    {
-      points: [
-        vertical.artifact,
-        point(vertical.artifact.x, horizontal.owner.y),
-        horizontal.owner
-      ],
-      artifactAxis: 'vertical',
-      ownerAxis: 'horizontal'
-    },
-    {
-      points: [
-        horizontal.artifact,
-        point(
-          (horizontal.artifact.x + horizontal.owner.x) / 2 + channelOffset,
-          horizontal.artifact.y
-        ),
-        point(
-          (horizontal.artifact.x + horizontal.owner.x) / 2 + channelOffset,
-          horizontal.owner.y
-        ),
-        horizontal.owner
-      ],
-      artifactAxis: 'horizontal',
-      ownerAxis: 'horizontal'
-    },
-    {
-      points: [
-        vertical.artifact,
-        point(
-          vertical.artifact.x,
-          (vertical.artifact.y + vertical.owner.y) / 2 + channelOffset
-        ),
-        point(
-          vertical.owner.x,
-          (vertical.artifact.y + vertical.owner.y) / 2 + channelOffset
-        ),
-        vertical.owner
-      ],
-      artifactAxis: 'vertical',
-      ownerAxis: 'vertical'
-    }
-  ].map(candidate => ({
-    ...candidate,
-    points: simplifyOrthogonalRoute(candidate.points)
-  })).filter(({ points, artifactAxis, ownerAxis }) => {
-    return routeHasDockDirections(points, artifactAxis, ownerAxis);
-  });
-
-  return candidates.map(({ points }, index) => {
-    const clear = routeAvoidsEndpointInteriors(
-      points,
-      artifactBounds,
-      ownerBounds
-    ) && pathIsClear(
-      points,
-      obstacles,
-      owner,
-      artifact,
-      [],
-      -ASSOCIATION_OBSTACLE_CLEARANCE
-    );
-
-    return {
-      points,
-      score: [
-        clear ? 0 : 1,
-        routeCrossings(points, routeSegments),
-        points.length - 2,
-        index,
-        routeLength(points)
-      ]
-    };
-  }).sort((a, b) => compareScores(a.score, b.score))[0].points;
-}
-
-function routeAvoidsEndpointInteriors(points, artifactBounds, ownerBounds) {
-  return toSegments(points).every(([ start, end ]) => {
-    return !segmentEntersRect(start, end, artifactBounds) &&
-      !segmentEntersRect(start, end, ownerBounds);
-  });
-}
-
-function routeHasDockDirections(points, artifactAxis, ownerAxis) {
-  const first = points[0];
-  const second = points[1];
-  const penultimate = points.at(-2);
-  const last = points.at(-1);
-  const firstAxis = first.y === second.y ? 'horizontal' : 'vertical';
-  const lastAxis = penultimate.y === last.y ? 'horizontal' : 'vertical';
-
-  return firstAxis === artifactAxis && lastAxis === ownerAxis;
-}
-
-function horizontalAssociationDocks(
-    artifactBounds,
-    ownerBounds,
-    connectionIndex,
-    connectionCount) {
-  const artifactLeft = artifactBounds.x + artifactBounds.width / 2 <
-    ownerBounds.x + ownerBounds.width / 2;
-  const offset = associationDockOffset(
-    ownerBounds.height,
-    artifactBounds.height,
-    connectionIndex,
-    connectionCount
-  );
-
-  return {
-    artifact: point(
-      artifactLeft
-        ? artifactBounds.x + artifactBounds.width
-        : artifactBounds.x,
-      artifactBounds.y + artifactBounds.height / 2 + offset
-    ),
-    owner: point(
-      artifactLeft
-        ? ownerBounds.x
-        : ownerBounds.x + ownerBounds.width,
-      ownerBounds.y + ownerBounds.height / 2 + offset
-    )
-  };
-}
-
-function verticalAssociationDocks(
-    artifactBounds,
-    ownerBounds,
-    connectionIndex,
-    connectionCount) {
-  const artifactAbove = artifactBounds.y + artifactBounds.height / 2 <
-    ownerBounds.y + ownerBounds.height / 2;
-  const offset = associationDockOffset(
-    ownerBounds.width,
-    artifactBounds.width,
-    connectionIndex,
-    connectionCount
-  );
-
-  return {
-    artifact: point(
-      artifactBounds.x + artifactBounds.width / 2 + offset,
-      artifactAbove
-        ? artifactBounds.y + artifactBounds.height
-        : artifactBounds.y
-    ),
-    owner: point(
-      ownerBounds.x + ownerBounds.width / 2 + offset,
-      artifactAbove
-        ? ownerBounds.y
-        : ownerBounds.y + ownerBounds.height
-    )
-  };
-}
-
-function simplifyOrthogonalRoute(points) {
-  const deduplicated = points.filter((candidate, index) => {
-    const previous = points[index - 1];
-
-    return !previous ||
-      candidate.x !== previous.x ||
-      candidate.y !== previous.y;
-  });
-
-  return deduplicated.filter((candidate, index) => {
-    const previous = deduplicated[index - 1];
-    const next = deduplicated[index + 1];
-
-    return !previous || !next || !(
-      (previous.x === candidate.x && candidate.x === next.x) ||
-      (previous.y === candidate.y && candidate.y === next.y)
-    );
-  });
-}
-
-export function artifactSizeCandidates(element) {
-  if (!is(element, 'bpmn:TextAnnotation')) {
-    return [ getDefaultSize(element) ];
-  }
-
-  const text = element.text || '';
-  const candidates = [];
-
-  for (
-    let width = ANNOTATION_MIN_WIDTH;
-    width <= ANNOTATION_MAX_WIDTH;
-    width += ANNOTATION_WIDTH_STEP
-  ) {
-    const lineCount = wrappedTextLineCount(text, width);
-    const height = Math.max(
-      ANNOTATION_MIN_HEIGHT,
-      lineCount * ANNOTATION_LINE_HEIGHT + 2 * ANNOTATION_PADDING
-    );
-
-    candidates.push({ width, height });
-  }
-
-  return candidates.sort((a, b) => {
-    return annotationSizePenalty(a) - annotationSizePenalty(b) ||
-      a.width * a.height - b.width * b.height ||
-      a.width - b.width;
-  });
-}
-
-function wrappedTextLineCount(text, width) {
-  const capacity = Math.max(
-    1,
-    Math.floor((width - 2 * ANNOTATION_PADDING) / ANNOTATION_CHARACTER_WIDTH)
-  );
-
-  return String(text || '').split(/\r?\n/).reduce((total, paragraph) => {
-    const words = paragraph.trim().split(/\s+/).filter(Boolean);
-
-    if (!words.length) {
-      return total + 1;
-    }
-
-    let lines = 1;
-    let lineLength = 0;
-
-    for (const word of words) {
-      const wordLength = word.length;
-
-      if (!lineLength) {
-        lines += Math.max(0, Math.ceil(wordLength / capacity) - 1);
-        lineLength = wordLength % capacity || Math.min(wordLength, capacity);
-      } else if (lineLength + 1 + wordLength <= capacity) {
-        lineLength += 1 + wordLength;
-      } else {
-        lines += Math.max(1, Math.ceil(wordLength / capacity));
-        lineLength = wordLength % capacity || Math.min(wordLength, capacity);
-      }
-    }
-
-    return total + lines;
-  }, 0);
-}
-
-function annotationSizePenalty(size) {
-  return Math.round(
-    Math.abs(size.width / size.height - ANNOTATION_TARGET_ASPECT_RATIO) *
-    ANNOTATION_ASPECT_RATIO_PENALTY_SCALE
-  );
-}
-
-export function findArtifactPlacement(
-    artifact,
-    {
-      ownerBounds,
-      references,
-      sizes,
-      obstacles,
-      routes,
-      occupied,
-      container,
-      processContainer,
-      extents,
-      annotationClearance = 0,
-      boundaryContainers = [],
-      avoidParticipantInterior = false,
-      preferParticipantSides = true,
-      participantInteriorPreference = 0
-    }) {
-  const preparedBoundaryContainers = boundaryContainers.map(({
-    rect,
-    containsOwner,
-    participant
-  }) => {
-    return {
-      containsOwner,
-      participant,
-      rect,
-      headerRect: participant && {
-        x: rect.x,
-        y: rect.y,
-        width: PARTICIPANT_HEADER_WIDTH,
-        height: rect.height
-      }
-    };
-  });
-  const referenceOwnerBounds = [ ...new Set(
-    references.map(reference => reference.ownerBounds).filter(Boolean)
-  ) ];
-  const hasMultipleOwners = referenceOwnerBounds.length > 1;
-  const localCandidates = ownerBounds
-    ? referenceOwnerBounds.flatMap(referenceBounds => {
-      return ownedArtifactCandidates(artifact, referenceBounds, sizes);
-    })
-    : unownedArtifactCandidates(sizes, extents);
-  const candidates = is(artifact, 'bpmn:TextAnnotation')
-    ? [
-      ...localCandidates,
-      ...participantExteriorArtifactCandidates(
-        sizes,
-        ownerBounds,
-        preparedBoundaryContainers,
-        annotationClearance
-      ),
-      ...outerScopeArtifactCandidates(sizes, extents)
-    ]
-    : localCandidates;
-  const obstacleBounds = obstacles.map(({ element, rect }) => {
-    return is(element, 'bpmn:BoundaryEvent')
-      ? artifactObstacleBounds(rect)
-      : rect;
-  });
-  const occupiedBounds = occupied.map(({ element, rect, annotationClearance }) => {
-    return artifactClearanceBounds(element, rect, annotationClearance);
-  });
-  const routeSegments = routes.flatMap(route => toSegments(route.points));
-  const dataArtifact = is(artifact, 'bpmn:DataObjectReference') ||
-    is(artifact, 'bpmn:DataStoreReference');
-  const requiresContainment = is(artifact, 'bpmn:DataObjectReference');
-  const rankedCandidates = candidates.map(candidate => {
-    const routedReferences = references.map(({
-      association,
-      owner,
-      ownerBounds,
-      ownerConnectionIndex,
-      ownerConnectionCount
-    }) => {
-      const points = ownerBounds && associationConnection(
-        association,
-        owner,
-        ownerBounds,
-        artifact,
-        candidate.rect,
-        ownerConnectionIndex,
-        ownerConnectionCount,
-        obstacles,
-        routeSegments
-      );
-      return points && { owner, points };
-    }).filter(Boolean);
-    const directLength = dataArtifact
-      ? ownerBalancedRouteLength(routedReferences)
-      : routedReferences.reduce((total, { points }) => {
-        return total + routeLength(points);
-      }, 0);
-    const associationCrossingCount = routedReferences.reduce((total, { points }) => {
-      return total + routeCrossings(points, routeSegments);
-    }, 0);
-    const associationBendCount = dataArtifact
-      ? routedReferences.reduce((total, { points }) => {
-        return total + Math.max(0, points.length - 2);
-      }, 0)
-      : 0;
-    const alignedOwnerCount = dataArtifact
-      ? referenceOwnerBounds.filter(referenceBounds => {
-        return artifactIsAligned(candidate.rect, referenceBounds);
-      }).length
-      : 0;
-    const missesOwnerAlignment = dataArtifact &&
-      referenceOwnerBounds.length > 0 &&
-      alignedOwnerCount === 0 ? 1 : 0;
-    const congestionPenalty = hasMultipleOwners
-      ? artifactCongestionPenalty(candidate.rect, obstacleBounds)
-      : 0;
-    const congestionViolation = congestionPenalty > 0 ? 1 : 0;
-    const participantInteriorPenalty = (
-      avoidParticipantInterior ||
-      participantInteriorPreference
-    ) &&
-      is(artifact, 'bpmn:TextAnnotation') &&
-      preparedBoundaryContainers.some(({
-        rect,
-        containsOwner,
-        participant
-      }) => {
-        return participant &&
-          containsOwner &&
-          artifactFitsContainer(candidate.rect, rect, 0);
-      }) ? 1 : 0;
-    const participantVerticalSidePenalty = preferParticipantSides &&
-      is(artifact, 'bpmn:TextAnnotation') &&
-      preparedBoundaryContainers.some(({
-        rect,
-        containsOwner,
-        participant
-      }) => {
-        return participant &&
-          containsOwner &&
-          (
-            candidate.rect.y + candidate.rect.height <= rect.y ||
-            candidate.rect.y >= rect.y + rect.height
-          );
-      }) ? 1 : 0;
-    const lowerBound = createArtifactPlacementScore({
-      participantInteriorViolation:
-        avoidParticipantInterior ? participantInteriorPenalty : 0,
-      dataCrossings: dataArtifact ? associationCrossingCount : 0,
-      congestionViolation,
-      missesOwnerAlignment,
-      alignedOwnerReward: -alignedOwnerCount,
-      associationBends: associationBendCount,
-      weightedLength: directLength +
-        congestionPenalty +
-        participantInteriorPenalty * participantInteriorPreference +
-        participantVerticalSidePenalty * 4 * VERTICAL_GAP,
-      associationLength: directLength,
-      annotationSize: is(artifact, 'bpmn:TextAnnotation')
-        ? annotationSizePenalty(candidate.rect)
-        : 0,
-      annotationCrossings: dataArtifact ? 0 : associationCrossingCount,
-      containmentViolation: 0,
-      expansion: artifactExpansion(candidate.rect, extents),
-      sideRank: candidate.sideRank,
-      offset: Math.abs(candidate.offset),
-      gap: candidate.gap,
-      y: candidate.rect.y,
-      x: candidate.rect.x
-    });
-
-    return {
-      candidate,
-      lowerBound,
-      alignedOwnerCount,
-      associationBendCount,
-      associationCrossingCount,
-      congestionPenalty,
-      congestionViolation,
-      missesOwnerAlignment,
-      participantInteriorPenalty,
-      participantVerticalSidePenalty,
-      routedReferences
-    };
-  }).sort((a, b) => {
-    return compareArtifactPlacementScores(a.lowerBound, b.lowerBound);
-  });
-
-  let best;
-
-  for (const {
-    candidate,
-    lowerBound,
-    alignedOwnerCount,
-    associationBendCount,
-    associationCrossingCount,
-    congestionPenalty,
-    congestionViolation,
-    missesOwnerAlignment,
-    participantInteriorPenalty,
-    participantVerticalSidePenalty,
-    routedReferences
-  } of rankedCandidates) {
-    if (best && compareArtifactPlacementScores(lowerBound, best.score) >= 0) {
-      break;
-    }
-
-    const candidateBounds = artifactClearanceBounds(
-      artifact,
-      candidate.rect,
-      annotationClearance
-    );
-    const fitsContainer = artifactFitsContainer(candidateBounds, container, 0);
-    const fitsProcessContainer = artifactFitsContainer(
-      candidateBounds,
-      processContainer,
-      0
-    );
-    const straddlesContainer = container &&
-      rectanglesOverlap(candidateBounds, container) &&
-      !fitsContainer;
-    const straddlesProcessContainer = processContainer &&
-      rectanglesOverlap(candidateBounds, processContainer) &&
-      !fitsProcessContainer;
-    const violatesContainerBoundary = preparedBoundaryContainers.some(({
-      rect,
-      containsOwner,
-      participant,
-      headerRect
-    }) => {
-      if (participant && rectanglesOverlap(candidateBounds, headerRect)) {
-        return true;
-      }
-
-      if (!rectanglesOverlap(candidateBounds, rect)) {
-        return false;
-      }
-
-      return !containsOwner || !artifactFitsContainer(candidateBounds, rect, 0);
-    });
-    const overlapsParticipantHeader = container &&
-      candidateBounds.y < container.y + container.height &&
-      candidateBounds.y + candidateBounds.height > container.y &&
-      candidateBounds.x < container.x &&
-      candidateBounds.x + candidateBounds.width >
-        container.x - PARTICIPANT_HEADER_WIDTH;
-
-    if ((requiresContainment && !fitsContainer) ||
-        straddlesContainer ||
-        straddlesProcessContainer ||
-        violatesContainerBoundary ||
-        overlapsParticipantHeader ||
-        obstacleBounds.some(obstacle => {
-          return rectanglesOverlap(candidateBounds, obstacle);
-        }) ||
-        occupiedBounds.some(occupied => {
-          return rectanglesOverlap(candidateBounds, occupied);
-        }) ||
-        artifactIntersectsRoutes(candidate.rect, routeSegments)) {
-      continue;
-    }
-
-    const directPathsClear = routedReferences.every(({ owner, points }) => {
-      return pathIsClear(
-        points,
-        obstacles,
-        owner,
-        artifact,
-        []
-      );
-    });
-
-    if (!directPathsClear) {
-      continue;
-    }
-
-    const associationRoutes = routedReferences.map(({ points }) => points);
-    const associationLength = dataArtifact
-      ? ownerBalancedRouteLength(routedReferences)
-      : associationRoutes.reduce((total, points) => {
-        return total + routeLength(points);
-      }, 0);
-    const nonStraightPenalty = is(artifact, 'bpmn:TextAnnotation')
-      ? associationRoutes.reduce((total, points) => {
-        return total + (
-          isStraightOrthogonalRoute(points)
-            ? 0
-            : NON_STRAIGHT_ARTIFACT_ASSOCIATION_PENALTY
-        );
-      }, 0)
-      : 0;
-    const score = createArtifactPlacementScore({
-      participantInteriorViolation:
-        avoidParticipantInterior ? participantInteriorPenalty : 0,
-      dataCrossings: dataArtifact ? associationCrossingCount : 0,
-      congestionViolation,
-      missesOwnerAlignment,
-      alignedOwnerReward: -alignedOwnerCount,
-      associationBends: associationBendCount,
-      weightedLength: associationLength +
-        congestionPenalty +
-        nonStraightPenalty +
-        participantInteriorPenalty * participantInteriorPreference +
-        participantVerticalSidePenalty * 4 * VERTICAL_GAP,
-      associationLength,
-      annotationSize: is(artifact, 'bpmn:TextAnnotation')
-        ? annotationSizePenalty(candidate.rect)
-        : 0,
-      annotationCrossings: dataArtifact ? 0 : associationCrossingCount,
-      containmentViolation: !is(artifact, 'bpmn:TextAnnotation') && (
-        (container && !fitsContainer) ||
-        (processContainer && !fitsProcessContainer)
-      ) ? 1 : 0,
-      expansion: artifactExpansion(candidate.rect, extents),
-      sideRank: candidate.sideRank,
-      offset: Math.abs(candidate.offset),
-      gap: candidate.gap,
-      y: candidate.rect.y,
-      x: candidate.rect.x
-    });
-
-    if (!best || compareArtifactPlacementScores(score, best.score) < 0) {
-      best = { candidate: candidate.rect, score };
-    }
-  }
-
-  if (best) {
-    return best.candidate;
-  }
-
-  throw new LayoutError(
-    'ARTIFACT_PLACEMENT_FAILED',
-    artifact.id,
-    `No collision-free artifact position could be found (${artifact.id}).`
-  );
-}
-
-export function compareArtifactPlacementScores(a, b) {
-  return compareScores(a, b);
-}
-
-export function createArtifactPlacementScore({
-  participantInteriorViolation,
-  dataCrossings,
-  congestionViolation,
-  missesOwnerAlignment,
-  alignedOwnerReward,
-  associationBends,
-  weightedLength,
-  associationLength,
-  annotationSize,
-  annotationCrossings,
-  containmentViolation,
-  expansion,
-  sideRank,
-  offset,
-  gap,
-  y,
-  x
-}) {
-  return [
-    participantInteriorViolation,
-    dataCrossings,
-    congestionViolation,
-    missesOwnerAlignment,
-    alignedOwnerReward,
-    associationBends,
-    weightedLength,
-    associationLength,
-    annotationSize,
-    annotationCrossings,
-    containmentViolation,
-    expansion,
-    sideRank,
-    offset,
-    gap,
-    y,
-    x
-  ];
-}
-
-function artifactIsAligned(artifact, owner) {
-  return artifact.x + artifact.width / 2 === owner.x + owner.width / 2 ||
-    artifact.y + artifact.height / 2 === owner.y + owner.height / 2;
-}
-
-function ownerBalancedRouteLength(routedReferences) {
-  const ownerRoutes = new Map();
-
-  for (const { owner, points } of routedReferences) {
-    const lengths = ownerRoutes.get(owner) || [];
-
-    lengths.push(routeLength(points));
-    ownerRoutes.set(owner, lengths);
-  }
-
-  return [ ...ownerRoutes.values() ].reduce((total, lengths) => {
-    return total + lengths.reduce((sum, length) => sum + length, 0) / lengths.length;
-  }, 0);
-}
-
-function artifactCongestionPenalty(rect, obstacles) {
-  const nearestDistance = obstacles.reduce((nearest, obstacle) => {
-    return Math.min(nearest, rectangleDistance(rect, obstacle));
-  }, Infinity);
-
-  return Math.max(0, 2 * ROUTING_MARGIN - nearestDistance);
-}
-
-function rectangleDistance(a, b) {
-  const horizontal = Math.max(a.x - b.x - b.width, b.x - a.x - a.width, 0);
-  const vertical = Math.max(a.y - b.y - b.height, b.y - a.y - a.height, 0);
-
-  return Math.hypot(horizontal, vertical);
-}
-
-function ownedArtifactCandidates(artifact, owner, sizes) {
-  const candidates = [];
-  const preferredSides = is(artifact, 'bpmn:DataObjectReference')
-    ? [ 'below', 'right', 'left', 'above' ]
-    : [ 'above', 'left', 'right', 'below' ];
-  const offsets = [ 0 ];
-
-  for (let distance = ROUTING_MARGIN; distance <= MAX_ARTIFACT_SEARCH_OFFSET; distance += ROUTING_MARGIN) {
-    offsets.push(distance, -distance);
-  }
-
-  for (const size of sizes) {
-    for (let sideRank = 0; sideRank < preferredSides.length; sideRank++) {
-      const side = preferredSides[sideRank];
-
-      for (let gap = ROUTING_MARGIN; gap <= MAX_ARTIFACT_GAP_STEPS * ROUTING_MARGIN; gap += ROUTING_MARGIN) {
-        for (const offset of offsets) {
-          candidates.push({
-            rect: artifactBoundsAt(owner, size, side, gap, offset),
-            sideRank,
-            offset,
-            gap
-          });
-        }
-      }
-    }
-  }
-
-  return candidates;
-}
-
-function artifactBoundsAt(owner, size, side, gap, offset) {
-  if (side === 'above') {
-    return bounds(
-      owner.x + (owner.width - size.width) / 2 + offset,
-      owner.y - gap - size.height,
-      size.width,
-      size.height
-    );
-  }
-  if (side === 'below') {
-    return bounds(
-      owner.x + (owner.width - size.width) / 2 + offset,
-      owner.y + owner.height + gap,
-      size.width,
-      size.height
-    );
-  }
-  if (side === 'left') {
-    return bounds(
-      owner.x - gap - size.width,
-      owner.y + (owner.height - size.height) / 2 + offset,
-      size.width,
-      size.height
-    );
-  }
-
-  return bounds(
-    owner.x + owner.width + gap,
-    owner.y + (owner.height - size.height) / 2 + offset,
-    size.width,
-    size.height
-  );
-}
-
-function artifactClearanceBounds(artifact, rect, clearance) {
-  if (!clearance || !is(artifact, 'bpmn:TextAnnotation')) {
-    return rect;
-  }
-
-  return bounds(
-    rect.x - clearance,
-    rect.y - clearance,
-    rect.width + 2 * clearance,
-    rect.height + 2 * clearance
-  );
-}
-
-function unownedArtifactCandidates(sizes, extents) {
-  const candidates = [];
-
-  for (const size of sizes) {
-    for (let offset = 0; offset <= extents.width + size.width; offset += ROUTING_MARGIN) {
-      candidates.push({
-        rect: bounds(
-          extents.minX + offset,
-          extents.minY - VERTICAL_GAP - size.height,
-          size.width,
-          size.height
-        ),
-        sideRank: 0,
-        offset,
-        gap: VERTICAL_GAP
-      });
-    }
-  }
-
-  return candidates;
-}
-
-function participantExteriorArtifactCandidates(
-    sizes,
-    ownerBounds,
-    boundaryContainers,
-    annotationClearance) {
-  if (!ownerBounds) {
-    return [];
-  }
-
-  const participants = boundaryContainers.filter(({
-    containsOwner,
-    participant
-  }) => {
-    return containsOwner && participant;
-  });
-
-  return participants.flatMap(({ rect }) => {
-    return sizes.flatMap(size => {
-      const x = ownerBounds.x + (ownerBounds.width - size.width) / 2;
-      const gap = Math.max(ROUTING_MARGIN, annotationClearance);
-
-      return [
-        {
-          rect: bounds(
-            x,
-            rect.y - gap - size.height,
-            size.width,
-            size.height
-          ),
-          sideRank: 4,
-          offset: 0,
-          gap
-        },
-        {
-          rect: bounds(
-            x,
-            rect.y + rect.height + gap,
-            size.width,
-            size.height
-          ),
-          sideRank: 4,
-          offset: 0,
-          gap
-        }
-      ];
-    });
-  });
-}
-
-function outerScopeArtifactCandidates(sizes, extents) {
-  const candidates = [];
-
-  for (const size of sizes) {
-    const horizontalPositions = axisPositions(
-      extents.minX,
-      extents.maxX - size.width
-    );
-    const verticalPositions = axisPositions(
-      extents.minY,
-      extents.maxY - size.height
-    );
-
-    for (const x of horizontalPositions) {
-      candidates.push({
-        rect: bounds(
-          x,
-          extents.minY - ROUTING_MARGIN - size.height,
-          size.width,
-          size.height
-        ),
-        sideRank: 4,
-        offset: 0,
-        gap: ROUTING_MARGIN
-      });
-      candidates.push({
-        rect: bounds(
-          x,
-          extents.maxY + ROUTING_MARGIN,
-          size.width,
-          size.height
-        ),
-        sideRank: 4,
-        offset: 0,
-        gap: ROUTING_MARGIN
-      });
-    }
-
-    for (const y of verticalPositions) {
-      candidates.push({
-        rect: bounds(
-          extents.minX - ROUTING_MARGIN - size.width,
-          y,
-          size.width,
-          size.height
-        ),
-        sideRank: 4,
-        offset: 0,
-        gap: ROUTING_MARGIN
-      });
-      candidates.push({
-        rect: bounds(
-          extents.maxX + ROUTING_MARGIN,
-          y,
-          size.width,
-          size.height
-        ),
-        sideRank: 4,
-        offset: 0,
-        gap: ROUTING_MARGIN
-      });
-    }
-  }
-
-  return candidates;
-}
-
-function axisPositions(min, max) {
-  if (max <= min) {
-    return [ Math.round((min + max) / 2) ];
-  }
-
-  const positions = [];
-
-  for (let position = min; position <= max; position += ROUTING_MARGIN) {
-    positions.push(position);
-  }
-
-  if (positions.at(-1) !== max) {
-    positions.push(max);
-  }
-
-  return positions;
-}
-
-function artifactFitsContainer(candidate, container, margin = ROUTING_MARGIN) {
-  return !container ||
-    candidate.x >= container.x + margin &&
-    candidate.y >= container.y + margin &&
-    candidate.x + candidate.width <=
-      container.x + container.width - margin &&
-    candidate.y + candidate.height <=
-      container.y + container.height - margin;
-}
-
-function artifactObstacleBounds(rect) {
-  return {
-    x: rect.x - BOUNDARY_EVENT_ARTIFACT_CLEARANCE,
-    y: rect.y - BOUNDARY_EVENT_ARTIFACT_CLEARANCE,
-    width: rect.width + 2 * BOUNDARY_EVENT_ARTIFACT_CLEARANCE,
-    height: rect.height + 2 * BOUNDARY_EVENT_ARTIFACT_CLEARANCE
-  };
-}
-
-function artifactIntersectsRoutes(rect, routeSegments) {
-  const clearance = {
-    x: rect.x - ROUTING_MARGIN / 2,
-    y: rect.y - ROUTING_MARGIN / 2,
-    width: rect.width + ROUTING_MARGIN,
-    height: rect.height + ROUTING_MARGIN
-  };
-
-  return routeSegments.some(([ start, end ]) => {
-    return segmentEntersRect(start, end, clearance);
-  });
-}
-
-function routeCrossings(points, routeSegments) {
-  return toSegments(points).reduce((total, [ start, end ]) => {
-    return total + routeSegments.filter(([ a, b ]) => {
-      return segmentsProperlyCross(start, end, a, b);
-    }).length;
-  }, 0);
-}
-
-function artifactExpansion(rect, extents) {
-  return Math.max(0, extents.minX - rect.x) +
-    Math.max(0, extents.minY - rect.y) +
-    Math.max(0, rect.x + rect.width - extents.maxX) +
-    Math.max(0, rect.y + rect.height - extents.maxY);
-}
-
-function isStraightOrthogonalRoute(points) {
-  return points.every(({ x }) => x === points[0].x) ||
-    points.every(({ y }) => y === points[0].y);
-}
-
-export function routeArtifactAssociation(
-    association,
-    owner,
-    ownerBounds,
-    artifact,
-    artifactBounds,
-    connectionIndex,
-    connectionCount,
-    obstacles = [],
-    routes = []) {
-  return associationConnection(
-    association,
-    owner,
-    ownerBounds,
-    artifact,
-    artifactBounds,
-    connectionIndex,
-    connectionCount,
-    obstacles,
-    routes.flatMap(route => toSegments(route.points))
-  );
-}
